refactor(auth): introduce base handler for EveMapper authorization with streamlined access checks
